Move theme methods to helper class
